+++ models/AuthorizationCode.php
@@ -1,16 +1,14 @@
 <?php
 /**
  * @link https://github.com/borodulin/yii2-oauth2-server
- * @copyright Copyright (c) 2015 Andrey Borodulin
  * @license https://github.com/borodulin/yii2-oauth2-server/blob/master/LICENSE
  */
 
 namespace conquer\oauth2\models;
 
+use conquer\oauth2\OAuth2;
 use Yii;
-use conquer\oauth2\Exception;
 use yii\db\ActiveRecord;
-use yii\helpers\VarDumper;
 
 /**
  * This is the model class for table "oauth_authorization_code".
@@ -23,68 +21,50 @@ use yii\helpers\VarDumper;
  * @property string $scope
  *
  * @property Client $client
- * @property ActiveRecord $user
+ *
+ * @author Andrey Borodulin
  */
 class AuthorizationCode extends ActiveRecord
 {
     /**
-     * @inheritdoc
+     * @throws \yii\base\InvalidConfigException
      */
     public static function tableName()
     {
-        return '{{%oauth2_authorization_code}}';
+        return OAuth2::instance()->authorizationCodeTable;
     }
 
     /**
-     * @inheritdoc
+     * @return string|\yii\db\Connection
+     * @throws \yii\base\InvalidConfigException
      */
-    public function rules()
+    public static function getDb()
     {
-        return [
-            [['authorization_code', 'client_id', 'user_id', 'expires'], 'required'],
-            [['user_id', 'expires'], 'integer'],
-            [['scope'], 'string'],
-            [['authorization_code'], 'string', 'max' => 40],
-            [['client_id'], 'string', 'max' => 80],
-            [['redirect_uri'], 'url'],
-        ];
-    }
-
-    /**
-     * @inheritdoc
-     */
-    public function attributeLabels()
-    {
-        return [
-            'authorization_code' => 'Authorization Code',
-            'client_id' => 'Client ID',
-            'user_id' => 'User ID',
-            'redirect_uri' => 'Redirect Uri',
-            'expires' => 'Expires',
-            'scope' => 'Scopes',
-        ];
+        return OAuth2::instance()->db;
     }
 
     /**
      *
-     * @param array $params
-     * @throws Exception
+     * @param $clientId
+     * @param $userId
+     * @param $scope
      * @return AuthorizationCode
      * @throws \yii\base\Exception
+     * @throws \yii\base\InvalidConfigException
      */
-    public static function createAuthorizationCode(array $params)
+    public static function createAuthorizationCode($clientId, $userId, $scope)
     {
-        static::deleteAll(['<', 'expires', time()]);
-
-        $params['authorization_code'] = Yii::$app->security->generateRandomString(40);
-        $authCode = new static($params);
-
-        if ($authCode->save()) {
-            return $authCode;
-        } else {
-            Yii::error(__CLASS__ . ' validation error: ' . VarDumper::dumpAsString($authCode->errors));
+        if (OAuth2::instance()->clearOldTokens) {
+            static::deleteAll(['<', 'expires', time()]);
         }
-        throw new Exception('Unable to create authorization code', Exception::SERVER_ERROR);
+        $authCode = new static();
+        $authCode->authorization_code = Yii::$app->security->generateRandomString(40);
+        $authCode->expires = time() + OAuth2::instance()->authCodeLifetime;
+        $authCode->client_id = $clientId;
+        $authCode->user_id = $userId;
+        $authCode->scope = $scope;
+        $authCode->save(false);
+        return $authCode;
     }
 
     /**
@@ -92,18 +72,6 @@ class AuthorizationCode extends ActiveRecord
      */
     public function getClient()
     {
-        return $this->hasOne(Client::className(), ['client_id' => 'client_id']);
-    }
-
-    /**
-     * @return \yii\db\ActiveQuery
-     */
-    public function getUser()
-    {
-        $identity = isset(Yii::$app->user->identity) ? Yii::$app->user->identity : null;
-        if ($identity instanceof ActiveRecord) {
-            return $this->hasOne(get_class($identity), ['user_id' => $identity->primaryKey()]);
-        }
-        return null;
+        return $this->hasOne(Client::class, ['client_id' => 'client_id']);
     }
 }
